+++ CHANGELOG.md
@@ -7,3 +7,8 @@
 - Took out snake part
 - Started rewriting the game using OOP principles
 - Started working on start screen
+
+# Day 3
+- Added custome button class
+- Finish start menu layout
+- **TODO**: hook up buttons to run the game